Create LEDEN analyse baseline
adding file
"LEDEN analyse baseline".  To cover changing values of "herkomst lidmaatschap" in ERP a baseline of original values is kept in a .csv file (ideally this will change to a db table)?  These original values can then be used in the procedure of running the "LEDEN analyse" so results from different years can be compared.